==> comparison_Bryois_2022/step3_scatterplots_GTExBrain.Rmd <==
---
title: "Scatterplots of -log10(pvalue) for eGenes in Bryois 2022"
author: "Si Liu and Wei Sun"
date: "`r Sys.Date()`"
output: 
  html_document:
    theme: journal
    highlight: tango
    toc: true
    toc_depth: 3
    toc_float:
      collapsed: true
      smooth_scroll: false
    number_sections: false
  df_print: paged
---
```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
options(width = 80)
```

For the eGenes from Bryois et al. (2022), we compare -log10(p-value) from Bryois et al. versus  -log10(p-value) by CSeQTL from GTEx brain bulk tissues. 

# Setup

## Load R libraries
```{r load_libraries, warning = FALSE, message = FALSE}
library(data.table)
library(readxl)
library(stringr)
library(ggplot2)
library(ggpubr)
library(dplyr)
library(reshape2)

theme_set(theme_classic())

```

## Data directory and file names
```{r}
CSeQTL_dir = "../../CSeQTL/data2share/"
GTExBrain_eqtl_file = "GTExBrain_eqtls.rds"

Bryois_2022_data_dir = "../../CSeQTL/Bryois_2022/"
Bryois_eGene_folder  = "Bryois_eGenes_50kb/"
```

## Cell type names
```{r}
Bryois_cell_types = c("Astrocytes", "Endothelial cells", 
                      "Excitatory neurons",  "Inhibitory neurons",  
                      "Microglia", "Oligodendrocytes", 
                     "OPCs / COPs", "Pericytes" )

GTExBrain_cell_types = c("Astro", "Exc", "Inh", "Micro", "Oligo", "OPC" )
```

## Load the genes considered for each cell type by CSeQTL

```{r}
GTExBrain <- readRDS(paste0(CSeQTL_dir, GTExBrain_eqtl_file))
length(GTExBrain)
names(GTExBrain)

lapply(GTExBrain, dim)

GTExBrain = GTExBrain[["eqtl"]]
dim(GTExBrain)
head(GTExBrain)
round(colSums(is.na(GTExBrain))/nrow(GTExBrain), 3)

length(unique(GTExBrain$ENSG))
length(unique(GTExBrain$GENE))
length(unique(GTExBrain$SNP))

table(GTExBrain$CELLTYPE, useNA = "ifany")
unique(GTExBrain$CELLTYPE)

# remove version number for gene ensemblID
GTExBrain$ENSG_version = GTExBrain$ENSG
GTExBrain$ENSG = gsub("\\..*", "", GTExBrain$ENSG_version)


GTExBrain = GTExBrain[which((GTExBrain$trim==TRUE) & 
                              (GTExBrain$perm==FALSE) & 
                              (GTExBrain$MODEL=="cseqtl") & 
                              (!is.na(GTExBrain$PVAL))), ]

dim(GTExBrain)

n_con_CSeQTL_list    = list()
gene_con_CSeQTL_list = list()

for (j in 1:length(GTExBrain_cell_types)){
  
  cct = GTExBrain_cell_types[j]
  
  cct_GTExBrain = GTExBrain[which(GTExBrain$CELLTYPE == cct), ]
  
  cct_considered_CSeQTL = unique(cct_GTExBrain$ENSG)
  
  stopifnot(nrow(cct_GTExBrain) == length(cct_considered_CSeQTL))
  
  n_con_CSeQTL_list[[cct]] = length(cct_considered_CSeQTL)
  gene_con_CSeQTL_list[[cct]] = cct_considered_CSeQTL

}

unlist(n_con_CSeQTL_list)

```


## Generate scatter plots

```{r}

n_Bryois_eGene_kept_mat =  matrix(NA, ncol=length(GTExBrain_cell_types), 
                                     nrow=length(Bryois_cell_types))

pdf(width = 12, height = 6, "figures/step3_pvalue_scatterplots_GTExBrain.pdf")

for (i in 1:length(Bryois_cell_types)){
  
  Bryois_ct = Bryois_cell_types[i]
  Bryois_ct_string = sub(" ", "_", str_squish(sub("/", "", Bryois_ct)))
  
  Bryois_ct_eGenes_path = paste0(Bryois_2022_data_dir, Bryois_eGene_folder, 
                                 "Bryois_eGenes_50kb_", 
                                 Bryois_ct_string, ".csv")
  
  df_Bryois_ct_eGenes = read.csv(Bryois_ct_eGenes_path, header = TRUE)
  Bryois_ct_eGenes    = df_Bryois_ct_eGenes$ensembl
  
  p_ct_list = list()

  for (j in 1:length(GTExBrain_cell_types)){
    
    CSeQTL_ct = GTExBrain_cell_types[j]
    CSeQTL_ct_considered = gene_con_CSeQTL_list[[CSeQTL_ct]]
    
    Bryois_ct_eGenes_kept = intersect(Bryois_ct_eGenes, CSeQTL_ct_considered)
    
    n_Bryois_eGene_kept_mat[i, j] = length(Bryois_ct_eGenes_kept)
    
    # extract Bryois pvalue for kept eGenes from Bryois 2022
    
    mat1 = match(Bryois_ct_eGenes_kept, df_Bryois_ct_eGenes$ensembl)
    Bryois_eGenes_Bryois_pval = df_Bryois_ct_eGenes[mat1, ]$bpval
    
    # extract CSeQTL GTExBrain pvalue for kept eGenes from Bryois 2022
  
    GTExBrain_CSeQTL_ct = GTExBrain[which(GTExBrain$CELLTYPE == CSeQTL_ct), ]
    
    Bryois_eGenes_CSeQTL_pval = 
      GTExBrain_CSeQTL_ct[match(Bryois_ct_eGenes_kept, 
                                GTExBrain_CSeQTL_ct$ENSG),]$PVAL
    
    check = length(Bryois_ct_eGenes_kept) == 
      sum(!is.na(Bryois_eGenes_CSeQTL_pval))
    stopifnot(check)
    
    # scatter plots of -log10(pvalue)
    
    df_ct_plot_pvalues = data.frame(Bryois = Bryois_eGenes_Bryois_pval, 
                                    CSeQTL = Bryois_eGenes_CSeQTL_pval)
    summary(df_ct_plot_pvalues)
    
    cur_corr = 
      cor(-log10(df_ct_plot_pvalues$CSeQTL), 
          -log10(df_ct_plot_pvalues$Bryois), method = "spearman")
    
    p_ct_list[[j]] = ggplot(df_ct_plot_pvalues, 
                            aes(x = -log10(CSeQTL), y = -log10(Bryois))) + 
      geom_point(size = 1, alpha = 0.5) + 
      geom_abline(intercept = 0, slope = 1) + 
      xlab("-log10(CSeQTL)") + ylab("-log10(Bryois)") + 
      ggtitle(paste0("Bryois: ", Bryois_ct_string, "\nCSeQTL: ", CSeQTL_ct, 
                     "\nSpearman corr: ", signif(cur_corr,2)))
  }
  
  print(ggarrange(plotlist=p_ct_list, nrow = 2, ncol = 4))
  
}

dev.off()

```


## Print out sanity check matrices

```{r}
rownames(n_Bryois_eGene_kept_mat) = Bryois_cell_types
colnames(n_Bryois_eGene_kept_mat) = GTExBrain_cell_types

n_Bryois_eGene_kept_mat

```

# Session information
```{r}
gc()

sessionInfo()
```
