==> comparison_Bryois_2022/step7_compare_eGenes_GTEx_brain.Rmd <==
---
title: "Check the difference between sc-eQTL results and CSeQTL"
author: "Si Liu and Wei Sun"
date: "`r Sys.Date()`"
output: 
  html_document:
    theme: journal
    highlight: tango
    toc: true
    toc_depth: 3
    toc_float:
      collapsed: true
      smooth_scroll: false
    number_sections: false
  df_print: paged
---
```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
options(width = 80)
```

Check the difference between the eGenes from Bryois et al. (2022) and eGenes identified by CSeQTL from GTEx brain bulk RNA-seq data.

# Setup

## Load R libraries
```{r load_libraries, warning = FALSE, message = FALSE}
library(data.table)
library(readxl)
library(stringr)
library(ggplot2)
library(ggpubr)
library(dplyr)
library(reshape2)
theme_set(theme_classic())

params = list(cs_q_cutoff=5e-3, fold_cutoff=1)
cs_q_cutoff = params$cs_q_cutoff
fold_cutoff = params$fold_cutoff

cat("cs_q_cutoff: ",  cs_q_cutoff,  "\n")
cat("fold_cutoff: ",  fold_cutoff,  "\n")

config = sprintf("cs_q_%.0e_fold_%.1f", 
                 cs_q_cutoff, fold_cutoff)
config
```

## Data directory and file names

```{r}
CSeQTL_dir = "../../CSeQTL/data2share/"
GTExBrain_eqtl_file = "GTExBrain_eqtls.rds"

Bryois_2022_data_dir = "../../CSeQTL/Bryois_2022/"

Bryois_eGene_folder  = "Bryois_eGenes_50kb/"
Bryois_con_gene_file = "Bryois_gene_considered.rds"
```

# Read data/results and make comparison

## Read eGenes for comparison. 

```{r}
fnm = sprintf("results/step2_%s_eGenes.rds", config)
eGenes_overlap_list = readRDS(fnm)
sapply(eGenes_overlap_list, function(x){sapply(x, length)})
```

## Cell type names

```{r}
Bryois_cell_types = c("Astrocytes", "Excitatory neurons", 
                      "Inhibitory neurons",  "Microglia", 
                      "Oligodendrocytes", "OPCs / COPs")

GTEx_cts = c("Astro", "Exc", "Inh", "Micro", "Oligo", "OPC")

stitle <- function(pval){
  str1 = sprintf("KS test p-val=%.1e", pval)
  if(pval < 0.01){ str1 = paste(str1, "**") }
  else if(pval < 0.05){ str1 = paste(str1, "*") }
  str1
}

theme1 = theme(axis.text.x = element_text(angle = 30, vjust = 1, 
                                   size = 11, hjust = 1),
               axis.text.y = element_text(size = 11),
               axis.title.x = element_blank(),
               legend.position = "none")

```

## Read in Bryois 2022 results

```{r fig.width=9, fig.height=11, warning = FALSE}

plist = list()

for (i in 1:length(Bryois_cell_types)){
  
  Bryois_ct = Bryois_cell_types[i]
  Bryois_ct_string = sub(" ", "_", str_squish(sub("/", "", Bryois_ct)))
  
  ct_nm = GTEx_cts[i]
  
  Bryois_ct_eGenes_path = paste0(Bryois_2022_data_dir, 
                                 Bryois_eGene_folder, 
                                 "Bryois_eGenes_50kb_", 
                                 Bryois_ct_string, ".csv")
  
  df_Bryois_ct  = read.csv(Bryois_ct_eGenes_path, header = TRUE)
  
  mat_i = match(eGenes_overlap_list[[i]]$Bryois, df_Bryois_ct$ensembl)
  stopifnot(sum(is.na(mat_i)) == 0)
  
  df_Bryois_ct = df_Bryois_ct[mat_i,]
  df_Bryois_ct$CSeQTL = 
    df_Bryois_ct$ensembl %in% eGenes_overlap_list[[i]]$CSeQTL

  df_Bryois_ct$CSeQTL_str = "sc-eQTL only"
  df_Bryois_ct$CSeQTL_str[df_Bryois_ct$CSeQTL] = "sc-eQTL and CSeQTL"
  df_Bryois_ct$CSeQTL_str = factor(df_Bryois_ct$CSeQTL_str, 
                                   levels = c("sc-eQTL only", "sc-eQTL and CSeQTL"))
    
  ks1 = ks.test(log10(df_Bryois_ct$dist[!df_Bryois_ct$CSeQTL] + 1), 
                log10(df_Bryois_ct$dist[df_Bryois_ct$CSeQTL]+1))
  
  plist[[3*(i-1)+1]] = ggplot(df_Bryois_ct, 
                              aes(x=CSeQTL_str, y=log10(dist+1), color=CSeQTL)) + 
      geom_violin(trim=FALSE) + geom_boxplot(width=0.1) + 
    ggtitle(sprintf("%s, %d vs. %d", ct_nm, sum(!df_Bryois_ct$CSeQTL), 
                    sum(df_Bryois_ct$CSeQTL)), 
            subtitle=stitle(ks1$p.value)) + theme1 + 
    ylab("log10(distance + 1)")
  
  ks1 = ks.test(-log10(df_Bryois_ct$adj_p)[!df_Bryois_ct$CSeQTL], 
                -log10(df_Bryois_ct$adj_p)[df_Bryois_ct$CSeQTL])

  plist[[3*(i-1)+2]] = ggplot(df_Bryois_ct, aes(x=CSeQTL_str, y=-log10(adj_p), 
                                                color=CSeQTL)) + 
      geom_violin(trim=FALSE) + geom_boxplot(width=0.1) + 
    ggtitle(label="", subtitle = stitle(ks1$p.value)) + theme1 + 
    ylab("-log10(adjuted p-value)")

  ks1 = ks.test(abs(df_Bryois_ct$beta)[!df_Bryois_ct$CSeQTL], 
                abs(df_Bryois_ct$beta)[df_Bryois_ct$CSeQTL])

  plist[[3*(i-1)+3]] = ggplot(df_Bryois_ct, aes(x=CSeQTL_str, y=abs(beta), 
                                                color=CSeQTL)) + 
      geom_violin(trim=FALSE) + geom_boxplot(width=0.1) + 
    ggtitle(label="", subtitle = stitle(ks1$p.value)) + theme1 + 
    ylab("|beta|")

}

ggarrange(plotlist=plist[c(1:6,13:15)], nrow = 3, ncol = 3)

ggarrange(plotlist=plist[-c(1:6,13:15)], nrow = 3, ncol = 3)
```

## Read in CSeQTL results

```{r fig.width=9, fig.height=11, warning = FALSE}
GTExBrain <- readRDS(paste0(CSeQTL_dir, GTExBrain_eqtl_file))
length(GTExBrain)
names(GTExBrain)

lapply(GTExBrain, dim)

GTExBrain = GTExBrain[["eqtl"]]
dim(GTExBrain)
GTExBrain[1:2,]

table(GTExBrain$CELLTYPE, useNA = "ifany")

GTExBrain$ENSG_version = GTExBrain$ENSG
GTExBrain$ENSG = gsub("\\..*","", GTExBrain$ENSG_version)

GTExBrain = GTExBrain[which((GTExBrain$trim==TRUE) & 
                              (GTExBrain$perm==FALSE) & 
                              (GTExBrain$MODEL=="cseqtl") & 
                              (!is.na(GTExBrain$PVAL))), ]

plist = list()

for (i in 1:length(GTEx_cts)){
  
  cct = GTEx_cts[i]
  df_GTEx = GTExBrain[which(GTExBrain$CELLTYPE == cct), ]
  
  mat_i = match(eGenes_overlap_list[[i]]$CSeQTL, df_GTEx$ENSG)
  stopifnot(sum(is.na(mat_i)) == 0)
  
  df_GTEx = df_GTEx[mat_i,]
  df_GTEx$sc_eGene = df_GTEx$ENSG %in% eGenes_overlap_list[[i]]$Bryois
  table(df_GTEx$sc_eGene)
  
  df_GTEx$sc_str = "CSeQTL only"
  df_GTEx$sc_str[df_GTEx$sc_eGene] = "CSeQTL and sc-eQTL"
  df_GTEx$sc_str = factor(df_GTEx$sc_str, 
                          levels = c("CSeQTL only", "CSeQTL and sc-eQTL"))

  table(df_GTEx$sc_str)

  ks1 = ks.test(log10(df_GTEx$MUa[!df_GTEx$sc_eGene]), 
                log10(df_GTEx$MUa[df_GTEx$sc_eGene]))
  
  plist[[3*(i-1)+1]] = ggplot(df_GTEx, 
                              aes(x=sc_str, y=log10(MUa), 
                                  color=sc_str)) + 
      geom_violin(trim=FALSE) + geom_boxplot(width=0.1) + 
    ggtitle(sprintf("%s, %d vs. %d", cct, sum(!df_GTEx$sc_eGene), 
                    sum(df_GTEx$sc_eGene)), 
            subtitle=stitle(ks1$p.value)) + theme1 + 
    ylab("log10(mean expression of A allele)")
  
  df_GTEx$qvalue[which(df_GTEx$qvalue < 1e-50)] = 1e-50
  
  ks1 = ks.test(-log10(df_GTEx$qvalue)[!df_GTEx$sc_eGene], 
                -log10(df_GTEx$qvalue)[df_GTEx$sc_eGene])

  plist[[3*(i-1)+2]] = ggplot(df_GTEx, aes(x=sc_str, y=-log10(qvalue), 
                                           color=sc_str)) + 
      geom_violin(trim=FALSE) + geom_boxplot(width=0.1) + 
    ggtitle(label="", subtitle = stitle(ks1$p.value)) + theme1 + 
    ylab("-log10(q-value)")

  ks1 = ks.test(abs(log10(df_GTEx$ETA))[!df_GTEx$sc_eGene], 
                abs(log10(df_GTEx$ETA))[df_GTEx$sc_eGene]) 

  plist[[3*(i-1)+3]] = ggplot(df_GTEx, aes(x=sc_str, y=abs(log10(ETA)), 
                                           color=sc_str)) + 
      geom_violin(trim=FALSE) + geom_boxplot(width=0.1) + 
    ggtitle(label="", subtitle = stitle(ks1$p.value)) + theme1 + 
    ylab("|log10(ETA)|")

}

ggarrange(plotlist=plist[c(1:6,13:15)], nrow = 3, ncol = 3)

ggarrange(plotlist=plist[-c(1:6,13:15)], nrow = 3, ncol = 3)
```

# Session information
```{r}
gc()

sessionInfo()
```
