==> comparison_Bryois_2022/step8_compare_eGenes_CMC.Rmd <==
---
title: "Overlap between Bryois 2022 and CMC case by Cell Type"
author: "Si Liu and Wei Sun"
date: "`r Sys.Date()`"
output: 
  html_document:
    theme: journal
    highlight: tango
    toc: true
    toc_depth: 3
    toc_float:
      collapsed: true
      smooth_scroll: false
    number_sections: false
  df_print: paged
params:
  AF: SCZ
---
```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
options(width = 80)
```

Check the overlap between the eGenes from Bryois et al. (2022) and eGenes identified by CSeQTL from Common Mind Consortium (CMC) Schizophrenia patients. 


# Setup

## Load R libraries
```{r load_libraries, warning = FALSE, message = FALSE}
library(data.table)
library(readxl)
library(stringr)
library(ggplot2)
library(ggpubr)
library(dplyr)
library(reshape2)

theme_set(theme_classic())

par_list = list(cs_q_cutoff=5e-3, fold_cutoff=1)
cs_q_cutoff = par_list$cs_q_cutoff # q value cutoff
fold_cutoff = par_list$fold_cutoff # fold change cutoff
AF = params$AF                   # Analysis Filter, SCZ vs. Control 

cat("cs_q_cutoff: ",  cs_q_cutoff,  "\n")
cat("fold_cutoff: ",  fold_cutoff,  "\n")
cat("AF: ",  AF,  "\n")

config = sprintf("%s_cs_q_%.0e_fold_%.1f", 
                 AF, cs_q_cutoff, fold_cutoff)
config
```

## Data directory and file names
```{r}
CSeQTL_dir  = "../../CSeQTL/data2share/"
CSeQTL_file = "CMC_eqtls.rds"

Bryois_2022_data_dir = "../../CSeQTL/Bryois_2022/"

Bryois_eGene_folder  = "Bryois_eGenes_50kb/"
Bryois_con_gene_file = "Bryois_gene_considered.rds"
```

# Read data/results and make comparison

## Read eGenes for comparison. 

```{r}
fnm = sprintf("results/step4_%s_eGenes.rds", config)
eGenes_overlap_list = readRDS(fnm)
sapply(eGenes_overlap_list, function(x){sapply(x, length)})
```

## Cell type names

```{r}
Bryois_cell_types = c("Astrocytes", "Excitatory neurons", 
                      "Inhibitory neurons",  "Microglia", 
                      "Oligodendrocytes", "OPCs / COPs")

CMC_cts = c("Astro", "Exc", "Inh", "Micro", "Oligo", "OPC")
```

## Read in Bryois 2022 results

```{r fig.width=9, fig.height=9, warning = FALSE}

plist = list()

for (i in 1:length(Bryois_cell_types)){
  
  Bryois_ct = Bryois_cell_types[i]
  Bryois_ct_string = sub(" ", "_", str_squish(sub("/", "", Bryois_ct)))
  
  ct_nm = CMC_cts[i]

  Bryois_ct_eGenes_path = paste0(Bryois_2022_data_dir, 
                                 Bryois_eGene_folder, 
                                 "Bryois_eGenes_50kb_", 
                                 Bryois_ct_string, ".csv")
  
  df_Bryois_ct  = read.csv(Bryois_ct_eGenes_path, header = TRUE)
  
  mat_i = match(eGenes_overlap_list[[i]]$Bryois, df_Bryois_ct$ensembl)
  stopifnot(sum(is.na(mat_i)) == 0)

  df_Bryois_ct = df_Bryois_ct[mat_i,]
  df_Bryois_ct$CSeQTL = 
    df_Bryois_ct$ensembl %in% eGenes_overlap_list[[i]]$CSeQTL
  table(df_Bryois_ct$CSeQTL)
  
  ks1 = ks.test(log10(df_Bryois_ct$dist[!df_Bryois_ct$CSeQTL] + 1), 
                log10(df_Bryois_ct$dist[df_Bryois_ct$CSeQTL]+1))
  
  plist[[3*(i-1)+1]] = ggplot(df_Bryois_ct, 
                              aes(x=CSeQTL, y=log10(dist+1), color=CSeQTL)) + 
      geom_violin(trim=FALSE) + geom_boxplot(width=0.1) + 
    ggtitle(sprintf("%s, %d vs. %d", ct_nm, sum(!df_Bryois_ct$CSeQTL), 
                    sum(df_Bryois_ct$CSeQTL)), 
            subtitle=sprintf("KS test p-val=%.1e", ks1$p.value))
  
  ks1 = ks.test(-log10(df_Bryois_ct$adj_p)[!df_Bryois_ct$CSeQTL], 
                -log10(df_Bryois_ct$adj_p)[df_Bryois_ct$CSeQTL])

  plist[[3*(i-1)+2]] = ggplot(df_Bryois_ct, aes(x=CSeQTL, y=-log10(adj_p),
                                                color=CSeQTL)) + 
      geom_violin(trim=FALSE) + geom_boxplot(width=0.1) + 
    ggtitle(label="", subtitle = sprintf("KS test p-val=%.1e", ks1$p.value))

  ks1 = ks.test(abs(df_Bryois_ct$beta)[!df_Bryois_ct$CSeQTL], 
                abs(df_Bryois_ct$beta)[df_Bryois_ct$CSeQTL])

  plist[[3*(i-1)+3]] = ggplot(df_Bryois_ct, aes(x=CSeQTL, y=abs(beta), 
                                                color=CSeQTL)) + 
      geom_violin(trim=FALSE) + geom_boxplot(width=0.1) + 
    ggtitle(label="", subtitle = sprintf("KS test p-val=%.1e", ks1$p.value))

}

ggarrange(plotlist=plist[c(1:6,13:15)], nrow = 3, ncol = 3, 
          common.legend=TRUE, legend="top")

ggarrange(plotlist=plist[-c(1:6,13:15)], nrow = 3, ncol = 3, 
          common.legend=TRUE, legend="top")

```

## Read in CSeQTL results

```{r fig.width=9, fig.height=9, warning = FALSE}
CSeQTL <- readRDS(paste0(CSeQTL_dir, CSeQTL_file))
length(CSeQTL)
names(CSeQTL)

lapply(CSeQTL, dim)

CSeQTL = CSeQTL[["eqtl"]]
table(CSeQTL$AF)

CSeQTL = CSeQTL[which(CSeQTL$AF==AF),]
dim(CSeQTL)
CSeQTL[1:2,]

table(CSeQTL$CELLTYPE, useNA = "ifany")

CSeQTL$ENSG_version = CSeQTL$ENSG
CSeQTL$ENSG = gsub("\\..*","", CSeQTL$ENSG_version)

CSeQTL = CSeQTL[which((CSeQTL$trim==TRUE) & 
                        (CSeQTL$perm==FALSE) & 
                        (CSeQTL$MODEL=="cseqtl") & 
                        (!is.na(CSeQTL$PVAL))), ]

plist = list()

for (i in 1:length(CMC_cts)){
  
  cct = CMC_cts[i]
  df_CMC = CSeQTL[which(CSeQTL$CELLTYPE == cct), ]
  
  mat_i = match(eGenes_overlap_list[[i]]$CSeQTL, df_CMC$ENSG)
  stopifnot(sum(is.na(mat_i)) == 0)
  
  df_CMC = df_CMC[mat_i,]
  df_CMC$sc_eGene = df_CMC$ENSG %in% eGenes_overlap_list[[i]]$Bryois
  table(df_CMC$sc_eGene)
  
  ks1 = ks.test(log10(df_CMC$MUa[!df_CMC$sc_eGene]), 
                log10(df_CMC$MUa[df_CMC$sc_eGene]))
  
  plist[[3*(i-1)+1]] = ggplot(df_CMC, 
                              aes(x=sc_eGene, y=log10(MUa), color=sc_eGene)) + 
      geom_violin(trim=FALSE) + geom_boxplot(width=0.1) + 
    ggtitle(sprintf("%s, %d vs. %d", cct, sum(!df_CMC$sc_eGene), 
                    sum(df_CMC$sc_eGene)), 
            subtitle=sprintf("KS test p-val=%.1e", ks1$p.value))
  
  df_CMC$qvalue[which(df_CMC$qvalue < 1e-50)] = 1e-50
  
  ks1 = ks.test(-log10(df_CMC$qvalue)[!df_CMC$sc_eGene], 
                -log10(df_CMC$qvalue)[df_CMC$sc_eGene])

  plist[[3*(i-1)+2]] = ggplot(df_CMC, aes(x=sc_eGene, y=-log10(qvalue), 
                                          color=sc_eGene)) + 
      geom_violin(trim=FALSE) + geom_boxplot(width=0.1) + 
    ggtitle(label="", subtitle = sprintf("KS test p-val=%.1e", ks1$p.value))

  ks1 = ks.test(abs(log10(df_CMC$ETA))[!df_CMC$sc_eGene], 
                abs(log10(df_CMC$ETA))[df_CMC$sc_eGene])

  plist[[3*(i-1)+3]] = ggplot(df_CMC, aes(x=sc_eGene, y=abs(log10(ETA)), 
                                          color=sc_eGene)) + 
      geom_violin(trim=FALSE) + geom_boxplot(width=0.1) + 
    ggtitle(label="", subtitle = sprintf("KS test p-val=%.1e", ks1$p.value))

}

ggarrange(plotlist=plist[c(1:6,13:15)], nrow = 3, ncol = 3, 
          common.legend=TRUE, legend="top")

ggarrange(plotlist=plist[-c(1:6,13:15)], nrow = 3, ncol = 3, 
          common.legend=TRUE, legend="top")
```

# Session information
```{r}
gc()

sessionInfo()
```
