==> comparison_Bryois_2022/step9b_summarize_results.Rmd <==
---
title: "Summarize eQTL direction comparison"
author: "Wei Sun"
date: "`r Sys.Date()`"
output: 
  html_document:
    theme: journal
    highlight: tango
    toc: true
    toc_depth: 3
    toc_float:
      collapsed: true
      smooth_scroll: false
    number_sections: false
  df_print: paged
params:
  AF: CMC_SCZ
---
```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
options(width = 80)
```

Check the difference between the eGenes from Bryois et al. (2022) and eGenes identified by CSeQTL from GTEx brain data.

# Setup

## Load R libraries
```{r load_libraries, warning = FALSE, message = FALSE}
library(data.table)
library(readxl)
library(stringr)
library(ggplot2)
library(ggpubr)
library(dplyr)
library(reshape2)
theme_set(theme_classic())

AF = params$AF # Analysis Filter, SCZ vs. Control 
cat("AF: ",  AF,  "\n")

```

## read in results

```{r}
fnm = sprintf("results/step9_compare_eQTL_directions_%s.tsv", AF)
overlap_all = fread(fnm)
dim(overlap_all)
overlap_all[1:2,]

table(is.na(overlap_all$beta))

c1 = chisq.test(overlap_all$beta > 0, overlap_all$ETA > 1)
c1$p.value
c1$expected
c1$observed
sum(diag(c1$expected))/nrow(overlap_all)
sum(diag(c1$observed))/nrow(overlap_all)

w2kp = which(overlap_all$qvalue < 0.1 & overlap_all$pval < 0.01)
length(w2kp)
c1 = chisq.test(overlap_all$beta[w2kp] > 0, overlap_all$ETA[w2kp] > 1)
c1
c1$expected
c1$observed
sum(diag(c1$expected))/length(w2kp)
sum(diag(c1$observed))/length(w2kp)

w2kp = which(overlap_all$qvalue < 0.005 & overlap_all$pval < 0.01)
length(w2kp)
c1 = chisq.test(overlap_all$beta[w2kp] > 0, overlap_all$ETA[w2kp] > 1)
c1
c1$expected
c1$observed
sum(diag(c1$expected))/length(w2kp)
sum(diag(c1$observed))/length(w2kp)

```

## Summarize overlap for each cell type separately

```{r}
table(overlap_all$cell_type)

signal1 = overlap_all$qvalue < 0.1 & overlap_all$pval < 0.01
table(overlap_all$cell_type, signal1)

signal2 = overlap_all$qvalue < 0.005 & overlap_all$pval < 0.01
table(overlap_all$cell_type, signal2)

cts = unique(overlap_all$cell_type)
length(cts)

v1 = rep(NA, 2*length(cts))
res = data.frame(cell_type = rep(cts, each=2), 
                 qvalue = rep(c(0.1, 0.005), times=length(cts)),
                 neQTL = v1, expected=v1, observed=v1, chisq_pval=v1)

for(i in 1:nrow(res)){
  ct1  = res$cell_type[i]
  qval = res$qvalue[i]
  
  w2kp = overlap_all$qvalue < qval & overlap_all$pval < 0.01
  w2kp = which(w2kp & overlap_all$cell_type == ct1)
  
  c1 = chisq.test(overlap_all$beta[w2kp] > 0, overlap_all$ETA[w2kp] > 1)
  
  res$neQTL[i]    = length(w2kp)
  res$expected[i] = round(sum(diag(c1$expected))/length(w2kp),2)
  res$observed[i] = round(sum(diag(c1$observed))/length(w2kp),2)
  res$chisq_pval[i] = signif(c1$p.value,2)
}

res
```

## plot
```{r fig.width=5.5, fig.height=3}
theme1 = theme(axis.text.x = element_text(angle = 60, vjust = 1, 
                                   size = 11, hjust = 1),
               axis.text.y = element_text(size = 11),
               legend.text=element_text(size=11))

res$qvalue = as.factor(res$qvalue)
p <- ggplot(res, aes(x=cell_type, y=observed, fill=qvalue)) +
  geom_bar(stat="identity", position=position_dodge()) + 
  ylab("prop. consistent eQTL direction") + xlab("")
p + scale_fill_brewer(palette="Dark2") + theme1

```

## Save results
```{r}
fnm = sprintf("results/step9b_eQTL_directions_summary_%s.tsv", AF)
fwrite(res, file=fnm, sep="\t")
```

# Session information
```{r}
gc()

sessionInfo()
```
